Move:enumやExceptionを別ファイルに移動
