+++ lib/service/github_repo_service.dart
@@ -1,7 +1,8 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
-import 'package:yumemi_codecheck_repo_search/generated/l10n.dart';
 import 'package:yumemi_codecheck_repo_search/model/repo_search_result.dart';
+
+export 'package:yumemi_codecheck_repo_search/common/exception/github_repo_service_exception.dart';
 
 part 'github_repo_service.g.dart';
 
@@ -33,41 +34,4 @@ class ExceptionalGitHubRepoService implements GitHubRepoService {
   }) async {
     throw _exception;
   }
-}
-
-typedef GRSException = GitHubRepoServiceException;
-
-class GitHubRepoServiceException implements Exception {
-  const GitHubRepoServiceException([this._message = '']);
-
-  final String _message;
-  String get message => _message;
-}
-
-class ValidationGRSException extends GRSException {
-  const ValidationGRSException();
-
-  @override
-  String get message => S.current.errorValidation;
-}
-
-class ServiceUnavailableGRSException extends GRSException {
-  const ServiceUnavailableGRSException();
-
-  @override
-  String get message => S.current.errorServiceUnavailable;
-}
-
-class NoInternetGRSException extends GRSException {
-  const NoInternetGRSException();
-
-  @override
-  String get message => S.current.errorNoInternet;
-}
-
-class UnexpectedGRSException extends GRSException {
-  const UnexpectedGRSException();
-
-  @override
-  String get message => S.current.errorUnexpected;
 }
